move stream handling functions to CameraUtils.js
